+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 ### Added
+- Added cluster support for `XADD`, `XDEL` and `XRANGE` (#1587)
 ### Changed
 - Refactor pipeline data writing depends on connection type (#1586)